refactor: add `components` folder
