+++ src/app/Components/ArtworkLists/views/CreateNewArtworkListView/components/InputRemainingCharactersLabel.tsx
